Changed the sending of DNS queries to a function "send_qry".
The function runs as a goroutine and sends a value to a channel.

The value sent to the channel is a struct with QPS (avg QPS) and RTT (avg RTT).

The user defined value for threads decides the number of send_qry goroutines to run.
At the end the channel is (supposed to be) iterated over and following values are taken.

QPS ( sum of average QPSs sent by the send_qry goroutines )
RTT ( Avegare of average RTTs sent by the send_qry goroutines )